ml/requirements.txt
# Core ML and Data Science
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning (if needed)
# tensorflow>=2.13.0  # Commented out for compatibility
# torch>=2.0.1        # Commented out for compatibility
# keras>=2.13.1       # Commented out for compatibility

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0

# API and Web
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0

# Model Serialization
joblib==1.3.2
pickle-mixin==1.0.2

# Visualization
plotly==5.15.0
bokeh==3.2.2

# Utilities
tqdm==4.65.0
requests==2.31.0
PyYAML==6.0.1
